frontend/requirements.txt
streamlit
psycopg2-binary 
streamlit-ace